chore: delete dead popup code; move shared helpers to Webcam/
webcamPopup.tsx and aiRatingBlock.ts were both retired by the RatingCard
slick redesign (PR #33) — the live map renders RatingCard via createRoot,
not the old HTML-string popup. Only the test files were still importing
them.

Moved formatModelLabel + SUNSET_DETECTION_THRESHOLD into a new file
co-located with their only consumer (AiRatingDisplay.tsx). The
aiRatingHelpers test suite covers what the old aiRatingBlock tests
covered for those exports.

Deleted:
- app/components/Map/lib/webcamPopup.tsx          (HTML popup generator)
- app/components/Map/lib/webcamPopup.test.ts      (tests for above)
- app/components/Map/lib/aiRatingBlock.ts         (HTML rating block)
- app/components/Map/lib/aiRatingBlock.test.ts    (tests for above)

Created:
- app/components/Webcam/aiRatingHelpers.ts        (shared helpers)
- app/components/Webcam/aiRatingHelpers.test.ts   (8 tests, all pass)

Updated:
- app/components/Webcam/AiRatingDisplay.tsx       (import path)
